requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.4.2
sqlalchemy==2.0.22
python-dotenv==1.0.0
pandas==2.1.1
plotly==5.17.0
aiohttp==3.8.5
apscheduler==3.10.4
streamlit==1.43.2
matplotlib==3.8.0
psycopg2-binary==2.9.9
protobuf==3.20.0
python-dateutil==2.8.2
setuptools==77.0.3
wheel==0.45.1 